technical chnages
Deleted multiple default layouts
Amalgamated layouts into one based on page attributes in YAML. Fluid
syntax in HTML is used to handle imagePosition and multiple images.

Played around a little more with
http://www.scarpetta.ch/vini%20del%20ticino/
It's not totally there yet, but highlighting and resizing works!

Regression tested rest of site to ensure changes did not result in any
bugs. - We are in a production environment now after all! =)